Work on handling room booking. Need to implement returning inserted record's PK id to finish.
